+++ src/pdfCropMargins/main_pdfCropMargins.py
@@ -51,6 +51,7 @@ import sys
 import os
 import shutil
 import time
+import readline # Makes prompts go to stdout rather than stderr.
 
 from . import __version__ # Get the version number from the __init__.py file.
 from .manpage_data import cmd_parser, DEFAULT_THRESHOLD_VALUE